Add packaging to program
