+++ src/main/java/commands/MarkTaskCmd.java
@@ -1,3 +1,5 @@
+package commands;
+
 public class MarkTaskCmd extends Command {
     private int taskIndex;
