+++ src/main/java/exceptions/GeneralException.java
